+++ packages/kit/src/extensions/design-export/index.ts
@@ -1,2 +1,4 @@
+export * from "./capability";
 export * from "./DesignExportExtension";
+export * from "./DesignExportCapabilityExtension";
 export * from "./commands";